pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kasukabe"
version = "1.0.0"
description = "Minecraft AI Building Studio — build from image or video"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "Pillow>=10.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "mcschematic>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["kasukabe*"]
